+++ CHANGES.md
@@ -4,6 +4,7 @@ Changes
 0.1.2 (unreleased)
 
 - Report parser errors through checks
+- Enforce strict keyword matching in parser
 
 
 0.1.1 (2019-06-30)