+++ CHANGES.md
@@ -1,7 +1,7 @@
 Changes
 =======
 
-0.1.2 (unreleased)
+0.1.2 (2019-07-06)
 
 - Report parser errors through checks
 - Enforce strict keyword matching in parser